bugfixes, eg plotting for linux envs

+++ copro/utils.py
@@ -106,7 +106,12 @@ def make_output_dir(config):
         for root, dirs, files in os.walk(out_dir):
             if config.getboolean('general', 'verbose'): print('DEBUG: remove files in folder {}'.format(os.path.abspath(root)))
             for fo in files:
-                os.remove(os.path.join(root, fo))
+                # print(fo)
+                if (fo == 'clf.pkl') or (fo =='XY.npy') or (fo == 'X.npy'):
+                    if config.getboolean('general', 'verbose'): print('DEBUG: sparing {}'.format(fo))
+                    pass
+                else:
+                    os.remove(os.path.join(root, fo))
 
     return out_dir
